test: move test utils to the root of the project
