--- frontend/__tests__/__utils__/default-data/index.ts
@@ -1,2 +1,0 @@
-export * from './custom-error';
-export * from './user-data';
