--- frontend/__tests__/__utils__/index.ts
@@ -1,2 +1,0 @@
-export * from './default-data';
-export * from './test-data';
