--- frontend/__tests__/components/users/domain/__mocks__/index.ts
@@ -1,1 +1,0 @@
-export * from './mock-user.repository';
