+++ frontend/jest.config.js
@@ -168,9 +168,7 @@ const config = {
   // ],
 
   // An array of regexp pattern strings that are matched against all test paths, matched tests are skipped
-  // A pattern to ignore index.ts inside __tests__ folder
-
-  testPathIgnorePatterns: ['/node_modules/', '/__utils__/', '/__mocks__/'],
+  // testPathIgnorePatterns: ['/node_modules/'],
 
   // The regexp pattern or array of patterns that Jest uses to detect test files
   // testRegex: [],
